pyproject.toml
[project]
name = "job"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apify-client>=2.3.0",
    "openai>=2.14.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
